Log container output in dev mode
Additionally save a file with the container output next to the
original file `[original].pdf.container_log`.

This is needed in order to increase visibility of non-critical
errors/warnings. Previously some issues went undetected because
if this [1].

Optionally, this output can be saved into a file
`<original_file>.container_log`, if the environment variable
DZ_LOG_CONTAINER is set. The last line is the success status of the doc.
Note: this file will be appended and not overwritten.

[1]: https://github.com/freedomofpress/dangerzone/issues/316